lower the image GC thresholds (#53)
Background: When new pods are being launched, new images are
pulled to the docker storage pool on the hosts.
old images (and layers) are kept for caching purposes even
when pods are deleted. Every 10 minutes, kubernetes node process
checks if the docker storage pool utilization is over
a threshold and deletes unused images if necessary. When the
total size of the pool is small (like when deploying on small
io-intensive flavors on cPouta) the pool may become full during
a build before cleaning kicks in.