remove extra openshift_hostname variable in inventory for nfs (#58)
Remove a an extraneous attribute and a invalid variable reference for nfs server in multimaster inventory template. This looks like it was introduced when openshift_hostname -attributes in inventory were added to provide human readable names for nodes. NFS server does not need it, because it does not act as an OpenShift node.