Set ansible_{ssh_user,become} for NFS servers
Previously the ansible_ssh_user and ansible_become variables were not
being set for NFS servers. This caused Ansible to be unable to connect
to these servers as the proper username was not being used. For other
servers the username is set in the OSEv3 group, but the NFS servers do
not belong to this group.

Set ansible_ssh_user to cloud-user and ansible_become to true in the NFS
server inventory group.