Make it clearer what each playbook does
Rename site.yml -> heat_site.yml since this is only used for stacks
provisioned with Heat. Rename configure.yml -> pre_install.yml as it is
actually only pre-installation configuration. This also makes sense when
there is also a post_install.yml. Amend README.md accordingly.

To conform with the naming conventions elsewhere in this repo, switch
from a dash to an underscore for the name heat_provision.yml.