Better reliability for OpenShift Heat provisioning
It looks like new routers in cPouta take a long time to start working.
It is thus better to use an existing router instead and connect that to
the cluster network create in the Heat stack. The name of this existing
router is provided as a parameter to the stack. Also change the default
CIDR of the cluster network to a /24 from a /16, because with an
existing router one has to avoid collissions with other networks
connected to the same router. Even then there are occasionally
connectivity issues with the new network. For these cases, make the
timeout for the initial connectivity check to the bastion host a
generous 15 minutes.

Improve SSH connectivity checking. The setup module is not ideal for
connection checking. It is better to use wait_for. First wait for the
bastion separately, then make sure the bastion has nmap-ncat installed
and finally ensure SSH connectivity to all cluster hosts.