mute errors in swap detection
Previously 'check for swap' tasks emit errors if the swap has not been
configured yet. The error is always ignored, but with Ansible 2.3
the summary at the end of the playbook shows these.

This patch will simplify swap creation and detection.
- separate playbook is eliminated
- all swap related tasks are put to a conditional block

Like previously, only one swap per host is supported.