Configuration for Cinder storage class
Add post-install tasks and a Jinja2 template for adding a Cinder storage
class in OpenShift.

Make it possible to configure a default storage quota for new projects
in OpenShift using a ResourceQuota object. The quota is based on a
variable and a default is provided for overriding the value. Set a
default of 50Gi to start with.