deploy prometheus and grafana behind a basic auth proxy
This commit adds support for conditionally deploying Prometheus with
Grafana for cluster wide performance monitoring. The deployment creates
a namespace called 'monitoring-infra' and deploys two pods:

- prometheus
  - has a side-car container for basic authentication
  - based on OpenShift 3.6 example by RedHat
  - visible as prometheus-monitoring-infra.CLUSTER_DOMAIN
- grafana
  - makes use of built in authentication
  - visible as grafana-monitoring-infra.CLUSTER_DOMAIN

After deployment, a Prometheus data source is added to Grafana using
the REST API.

The deployment is conditional and triggered with "deploy_monitoring"
variable. The deployment process configures authentication with a token
that must be provided with Ansible variable "monitoring_access_token".
Secret with keys for plain token and htaccess file is generated by
playbook and used by deployments.