Add .gitlab-ci.yml
Define a CI pipeline using .gitlab-ci.yml. Because we want to build
Docker images as a part of the pipeline, we use the "docker" Docker
image which is Alpine Linux with the Docker tools installed. We also use
the Docker-in-Docker (docker:dind) service to run nested Docker
containers. Initially the pipeline consists of building the poc-deployer
container, deploying an OpenShift environment and tearing down that
environment.

Some modifications to the included bash scripts were necessary both for
non-interactive use and for Docker-in-Docker support:

  - Add new -i option to run_deployment_container.bash. Using this
    option will add "-it" to the "docker run" command. Previously "-it"
    was always added.
  - Add new -P option to run_deployment_container.bash that allows
    passing of the Vault password from a file via an environment
    variable to the inner Docker container. This is a workaround for
    Docker-in-Docker that does not support bind mounts in a nice way.