run_playbook script: use exit code from Ansible
The run_playbook.bash script was always returning with an exit code of 0
since "set -e" is not enabled for that script and the last command is
"popd" instead of the ansible-playbook command whose exit code we are
actually interested in. Store the exit code from Ansible and use this to
exit from the script instead. This way failures in Ansible get passed on
and are picked up as failures when running in a GitLab runner.