Add Rocket.Chat template in openshift namespace
Add a new playbook that is called as part of the post-install phase to
create app templates and related objects that are stored in their own
apps subdirectory. The new playbook picks up any new files added under
the directory using with_fileglob.

Modify the upsert task slightly to accomodate the new with_fileglob
pattern: remove the .j2 from the src parameter to template so that the
filenames aren't always appended with .j2. Add the .j2 in calls to
upsert where necessary.

As a first use of this new playbook, add a Rocket.Chat app template in
the openshift namespace so that it is available for all end users to
create their own Rocket.Chat instances with.