upsert: make sure temp dir is actually deleted
Put the upsert task in a block where the deletion of the temp directory
always runs. Also, fix the deletion of the temp directory so that the
temp directory is actually deleted by making the file module delete the
value of the upsert_mktemp.stdout variable and not the non-existent path
"upsert_mktemp.stdout".