Fork public image at the start of site.yml
It looks like Heat stack updates break when the image used for creating
the virtual machines disappears (is deleted or made private). If using a
public image, the image may be disappear without notice due to updates.
To prevent this from happening, fork a given public image at the start
of deployment to be used as a base image for virtual machines that are
created by Heat. This image can then be updated in a controlled manner
and stack updates don't break because of this unexpectedly.