Install NRPE and Gearman for monitoring
The bastion host will be used as a Gearman proxy so install Gearman
there. Install NRPE everywhere else.

Since we are using ansible-galaxy to install roles necessary to install
NRPE and Gearman, add configuration that makes Ansible look for roles
under a new galaxy-roles/ directory in addition to the normal roles/
directory. Make sure roles installed in this way are ignored by amending
.gitignore. Make sure ansible-galaxy gets installed for CI runs.

Add a requirements.yml file that is fed to ansible-galaxy for role
installation. Lock all versions of all roles to avoid issues with role
updates.

We must also take installation of the third party roles into account in
our CI workflow, as these roles need to be installed before running any
playbooks. Amend before_script to address this.

Adding the Gearman server to the bastion host means that we need to also
allow port 4730 in addition to 22 on the bastion. This means passing the
bastion_allow_ports parameter to Heat stacks during build/update time.
Previously the default value of '22' was always used.

Allow access to port 5666 in the global default iptables template. All
hosts will be running the NRPE daemon same as all hosts run the SSH
daemon, so it's logical to manage both these ports in the same place -
i.e. the default firewall template.

Name resolution should also work on the bastion host, so make changes
for that as well.