Make all hosts use static IPs
The common cluster network is now created in a way where a separate
allocation pool is created for dynamically allocated addresses.
Addresses outside this pool can be assigned as static IP addresses. The
size of the allocation pool is fully configurable, though currently it
must be contiguous.

Introduce a scheme where the static IP of a host is determined based on
three variables:

 - Network prefix, which is the octets of the network's CIDR that don't
   change
 - Resource group identifier, which is a number that differentiates IP
   addresses for different resource groups
 - Index, which comes from the resource group itself and is an
   incrementing count for hosts in the resource group

To make this scheme work, start indexing for VM names in resource groups
from 1 using removal_policies. We also need to add a new task in the
environment context playbook for fetching variables based on install
type (currently either singlemaster or multimaster). This is required as
some of the networking data is available only in the install type
context to avoid repetition.

Also fix some bugs with determining when to create/update stacks.