Open NRPE port on OpenShift cluster nodes
Make sure port 5666 is open to the internal network on the cluster nodes
so that the bastion host that acts as Gearman proxy can send check jobs
to the NRPE daemons.

Add a test that ensures port 5666 is accessible from the bastion on all
hosts listed in /etc/hosts. In order to accomodate this test, it was
necessary to split the bats_wrapper script into a generic version and an
OpenShift version (what it used to be before). The OpenShift version
uses oc to create OpenShift objects, but this is something we don't
necessarily want to do for all tests.