Add configuration of GitHub OAuth provider
Add necessary roles and playbooks for configuring GitHub OAuth.
Currently this is done during the post-install phase as we want to be
able to make changes to these settings outside of openshift-ansible's
byo/config.yml playbook that can only run during initial installation.

Editing of the authentication configuration is done in a separate role
that is applied to masters. Use the yedit module available in
openshift-ansible to edit the master-config.yaml file. To access this
module, add the library directory of the relevant role inside
openshift-ansible to the library search path.

Another role is also added that contains handlers for restarting the
OpenShift master services. This role is added as a dependency for the
authentication role so that it can access these handlers.