Fixes for monitoring
Previously the monitoring connectivity test was not working properly
because nmap was not being installed on the bastion host. Install nmap
there to fix this.

Move the opening of the NRPE port from the openshift_node to the
cluster_common role. This port should be open on all hosts, so the
cluster_common role is more appropriate here. Since the "restart
iptables" handler is being called in so many places, move it to the
openshift_handlers role at the same time and set up dependencies in
roles that use this handler accordingly.

Run the cluster_common and base roles on the bastion separately instead
of running them on the bastion in the same play as on the other hosts.
With the added run of the "restart iptables" handler in cluster_common,
Ansible will get stuck if the handler runs on bastion and other hosts at
the same time. If the handler runs separately on bastion, Ansible won't
get stuck.