Disable kdump and run systemctl reset-failed
When using kdump, the crashkernel kernel option must be set. If the
option is not set, kdump will fail to start and will be listed as a
failed unit in the output of systemctl --failed. Since we don't use
kdump, disable it and run systemctl reset-failed to get rid of the
failed unit. If kdump was already disabled, do nothing.