Bastion: allow NRPE from mon server, not cluster
There should be no need for all hosts in the cluster to connect to the
bastion host's NRPE port, so don't allow this. Allow connections to the
NRPE port on the bastion only from the monitoring server instead. Change
the task that allows connections to Gearman to also allow connections to
NRPE and make sure iptables is notified on changes. This requires
setting a dependency on the openshift_handlers role in meta/main.yml.