Update NRPE role to remove hardcoded checks
The default configuration file for NRPE has some hardcoded checks that
can interfere with what the user actually wants to configure. Update
ansible-role-nrpe to a version that allows the removal of these
hardcoded checks. This fixes the check_load check's arguments so that
they are taken from the correct config file and not the default one.

See: https://github.com/CSCfi/ansible-role-nrpe/pull/7