Add test for pod health in default namespace
All the pods in the default OpenShift namespace should be in the state
"Running" after a cluster has been provisioned. There should be no
failed pods. Write a test that asserts this.

As the new test runs on the masters without the OpenShift project
wrapper, reorganize the code to accomodate for this. Also add a cleanup
for the test files as this was missing. Without this cleanup, running
these new tests against an old system will most likely cause errors. If
the structure of the tests changes later on, errors may occur because of
old leftover tests on the systems.