Playbook for backing up etcds and masters
Write a playbook that backs up data in an OpenShift cluster.

There is a role in openshift-ansible called etcd_common (or etcd with
versions > 3.6) that can be used to backup etcd data. To access this
role, add the roles from openshift-ansible to roles_path.

Keys and certs are also backed up from the etcd and the master hosts.

Multimaster and singlemaster installations are handled separately, as
etcd is installed in different hosts in these.

Backed up data is synced under a single directory on the bastion host.
In order for this to work, the cluster's private key needs to be stored
on the bastion instead of the masters. This also requires configuring
the cluster common parts on the bastion host first.