Compress etcd data and take all of /etc/origin
Data in an etcd cluster compresses quite well, so before transferring
backups to the bastion, make an archive out of the backup directory
structure. This makes it less likely that we will run out of storage on
the bastion and conserves network bandwidth.

It looks like more than just certificates are needed from the masters to
recover in a disaster scenario. Backup all of /etc/origin just in case
any of the other stuff under there is needed.

Also shorten some of the overly lengthy lines and reduce repetition.