Fix temp file cleanup for monitoring deployment
Previously the cleanup of the temp directory was being done on the
localhost and not on the host where the temp directory was created.
Also, the path for the task was not correctly set as there was a
variable missing curly brackets around it. Fix these issues so that the
temporary directory gets properly cleaned up.