Ensure /mnt is not mounted nor listed in fstab
This fixes a bug where mounting /mnt would fail after a reboot, causing
a failed systemd unit which would be reported in monitoring.

Change the cluster common role to make sure there are no default /mnt
mounts. In an OpenStack environment, this mount may sometimes be
automatically added because of cloud-init or the base image that is
used. Since we have no use for such mounts, ensure they don't exist.