add restore script and playbooks
This adds support for restoring cluster from backups. The control for
the process is full_restore.bash that invokes ansible and ansible-playbook
commands as needed. The rationale for using bash instead of an ansible
playbook is that we need to execute openshift-ansible playbooks selectively
by defining tags, that behaviour is not available in ansible playbook
include mechanisms.

Added:

- playbooks/restore subdirectory
- full_restore.bash - main script
- restore_etc.yml - playbook to restore /etc/ to etcd and master hosts
- restore_etcd_to_single_member.yml - playbook to restore etcd data and
  bring up a single member cluster
- a section on restore in README
Fixes:
- select_lb.bash: added check that PUBLIC_IP is set