Always check floating IP associations
Instead of doing floating IP associations in the same block as the Heat
stack updates, always check if floating IPs have been assigned. If for
some reason the Heat stacks are up but the floating IP of the master/LB
node is not assigned, then this situation would not have been fixed by
running Ansible. Only check for number of masters to determine whether
to assign a floating IP - not allow_heat_stack_update etc.