fix node stack update failure when site.yml is rerun
Add a condition to node stack updates so that it is only run if
base stack outputs are defined. This happens when
a) we are provisioning a fresh cluster
b) allow_heat_stack_update is defined

Also add comments to help identifying sm/mm blocks and fix indentation
on singlemaster block when conditions.