Fix certs for registry console/prometheus/grafana
Because the routes for Prometheus and Grafana are in the API domain, we
need to explicitly specify the certificates for these routes so that the
default app domain cert does not get used instead. Also add a redirect
policy for insecure traffic for these routes.

Also add a missing bool filter for replacement of the registry console
cert so that it is possible to set a replace variable for replacing the
registry console cert.