Run backups in the same jobs as deployments
Instead of using the separate backup jobs during normal deployments, run
the backup playbook as part of the deployment itself. Also change the
backup jobs so that they are only executed in scheduled pipelines.

The reason for this change is to avoid running the backup playbook
against a production system in cases where it is potentially harmful. If
the backups break something in oso-qa, there is a better chance that
this gets noticed before the backup code runs against production as the
production backups are only taken when the manual deployment task runs
or during scheduled pipelines.