Deploy default app from a repo instead of an image
When calling oc new-app to create the default www app (what is shown in
the main URL of a cluster), use a Git URL instead.

Take environment variables for the default app from a freely definable
dict instead of hardcoding a certain set of environment variables.

Add a test to ensure that the routes created for the default www app all
return normal HTTP status codes and don't show the "Application is not
available page".

Remove the source code for the old default www app as this is no longer
needed.