Add keepalived for multimaster installs
Make the LB pair in multimaster installations highly available by
installing keepalived. This requires some changes to the networking
setup of these virtual machines as well. The floating IP is no longer
attached to one of the machines but rather to a separate virtual IP port
that is then used by one of the LB machines. For this to work, the
allowed address pairs option is used for the instance ports in the LB
nodes. This change requires us to specify the subnet of the instance
port in Heat instead of the network.

Add a role for installing keepalived and a basic configuration file
template. Add a script to the role that checks for POC installation and
if POC is not installed, reduces the priority of that node. This is to
ensure that a replaced MASTER LB node does not take the virtual IP
before it is configured.