Single master: add master node to etcd group
In OpenShift 3.7, embedded etcd installations are no longer supported.
This means we must add the master node in single master installations
into the etcd group so that the OpenShift installer knows to install
etcd there in a separate container.

We must also modify the backup playbook so it does not try to backup an
embedded etcd on singlemaster installations but uses the same process
for both single- and multimaster install types.