Update documentation on replacing first etcd node
There is a patch in the release-3.6-csc branch of openshift-ansible that
allows you to set a variable called override_etcd_ca_host that will be
used to override the value of etcd_ca_host. However, it is also possible
to directly set etcd_ca_host without the need to have an intermediate
variable. Update the documentation to reflect this so we can drop the
patch from our fork of openshift-ansible.

Ansible variable precedence prefers extra vars set on the command line
over any other variable setting, so that's why setting -e etcd_ca_host
overrides what is set in the vars section of a play.