Instead of removing Heketi route, secure it
OpenShift 3.7 has a bug that means that the master controller container
doesn't have its DNS configured properly after installation. Restarting
the controller helps, but doing that is problematic for the Ansible
tasks that during post-install. This affects using the IP from the
Heketi storage Service object directly for accessing the Heketi API, as
this can't be resolved immediately after installation from the
controller container. We must use the Route created during installation
instead. However, we want to modify the Route first so that it is
secure.

The modified heketi-storage route has a whitelist applied to it and is
secured with TLS.