Prometheus+Grafana: Fixes for OpenShift 3.7
With Kubernetes 1.7, each node exposes a cAdvisor endpoint that reports
container metrics. Add a scrape config for this endpoint to Prometheus
so that we get metrics with OpenShift 3.7.

Also, the cluster filesystem status in the Kubernetes dashboard in
Grafana wasn't showing anything, so replace it with a Docker pool
storage availability status. This is more useful than, say, the
available space in all the root filesystems on the cluster nodes that
the previous cluster filesystem apparently tried to show.