Add scripts for LB replacement and polling
Based on the LB replacement documentation, add a one-off update script
that will replace load balancers. Also add a script for polling an
address to measure the downtime caused by running the LB replacement
script.

A new flag needs to be added to provisioning for the script to work. The
flag can be used to toggle checking for SSH connectivity after
provisioning. Initially servers whose ports are replaced will lose their
connectivity before they are rebuilt due to a MAC clash between the old
ports and the new ones.