Make it possible to extend Gluster by adding vols
Modify the GlusterFS Heat stack and its substacks so that it is
possible to add a configurable number of extra volumes that all have the
same configurable size and to add these to every member of the GlusterFS
cluster. Rename the vm_with_two_volumes stack as that name would no
longer be accurate.