Safety checks for GlusterFS provisioning
Assert that we are not trying to shrink the number of GlusterFS nodes or
volumes. These assertions can be disabled with a boolean if need be, but
they are done by default.

The new assertions require some changes to how Heat stack existence is
checked. We now get additional information about how the stacks were
previously provisioned which is used for the assertions. The output
remains the same in case the stacks don't exist, so existing
functionality is also preserved.

Add deletion policies for volumes attached to the GlusterFS cluster so
that they are not deleted by default when the stack is deleted or
updated. The deletion policy needs to be explicitly set to "Delete" for
devel/CI environments.