Ensure GlusterFS volume ordering with depends_on
The extension volumes for the GlusterFS stack did not have depends_on
set, so they would be created before the second volume that is the
default data volume for a GlusterFS node. This was causing some
unexpected behaviour in CI runs. Set depends_on: volume_attachment_2 on
the extension volume ResourceGroup so that the default data volume is
always created and attached before the extension volumes.