Docker: Do not always configure VG and DATA_SIZE
For docker-storage-setup, we don't necessarily want to set the VG and
DATA_SIZE options. These were previously hardcoded. This is OK when
using devicemapper as the driver, but for e.g. overlay2 we don't want to
set these options.

Modify the code so that the VG and DATA_SIZE options are only set when
the driver is set to devicemapper. Otherwise the corresponding Ansible
variables are set to an empty string. Additionally, modify the
docker_host role so that it won't set these options when their
corresponding Ansible variables are set to empty values. Also make
overlay2 the default storage driver in the docker_host role.

After these modifications we can configure both devicemapper and
overlay2 depending on what config variables we set for an environment.