Put vault pass writing commands into a script
The commands for storing the vault password were basically already in
script form in README.md, so make it into an actual script. Add some
printing for usability and ensure all clipboards are cleared after the
vault pass has been written to the RAM disk.

To avoid repetition, remove the commands from README.md and reference
the script in their place.