Split site.yml for OpenShift 3.7 and 3.9
The location of the installation playbooks changes in openshift-ansible
3.9, so we must make different versions of our playbooks depending on
whether we are installing 3.7 or 3.9. It looks like Ansible does not
fully support dynamically deciding which file to import at runtime based
on a variable, so we have to create separate playbooks for the different
OpenShift versions.

For backwards compatibility site.yml behaves the same way as before by
including the new site_3.7.yml file. For 3.9, a site_3.9.yml file is
included. The difference between the two site files is the install
playbook they include (install_3.7/3.9). The install playbook calls
whatever playbook(s) in openshift-ansible happen(s) to be the correct
one(s).