Upgrade deployment container Ansible to 2.4
Update what gets installed in requirements.txt for the deployment
container so that we install the latest point release of Ansible 2.4
instead of Ansible 2.3. The packages to upgrade/install were determined
by installing just Ansible in its own virtualenv, running pip freeze and
comparing the resulting list with requirements.txt.