Use version_compare instead of == for version comp
Update how handlers and the resolv.conf fix for 3.7 are called based on
version specific conditionals. Instead of a simple equal checks, use the
version_compare version comparison filter:

http://docs.ansible.com/ansible/latest/user_guide/playbooks_tests.html#version-comparison

This has the added benefit of not having to update the conditionals
every time there is an OpenShift version update. We simply do some
things when version >= 3.7 and some other things when version equals
3.6.