Nagios check for OpenShift deployments
Add a Nagios check that creates a DeploymentConfig with a basic web
server serving content, the necessary API objects to expose this web
server to the world and a polling check to ensure the correct content is
visible behind the URL exposed.

The Nagios check runs via a wrapper that creates a virtualenv with the
required packages and logs in to OpenShift. For this we need credentials
available on the bastion host where the check is intended to run. Move
the copying of test user credentials from the test playbooks over to the
bastion role while removing the task to remove the test credentials
after testing is done.

Use the Nagios check in two new BATS tests so that the test set that
runs after OpenShift deployments has an end-to-end functionality test
for deployments.