Test more than just default namespace health
Instead of only checking for the health of the default namespace, check
a selection of other system namespaces as well. Generalize the test code
for checking namespace health so that it takes a namespace name instead
of being hardcoded to only check for the default namespace.

Also fix some issues that prevented us from getting nice output on
failed tests. Add a "rescue" section to the test blocks that prints test
output in human readable form if there are errors in the tests. Use
import_playbook instead of include to fix Ansible deprecation warnings.

These new tests are meant to ensure that all system components are
healthy instead of just the ones in the "default" namespace. The
motivation to work on this originally came from fixing an issue where
the metrics image tag was too specific and thus the metrics pods were
not starting due to a missing image.