Lock OS minor version and disable autoupdates
Ensure that updates are not automatically made from one OS minor version
to another. These should be handled in a more controlled fashion. Remove
the default .repo config file and replace it with modified .repo files
that lock down the minor version.

For this to work, we need to reorder the pre-install configuration steps a
bit:
 - Instead of only installing nmap-ncat during provisioning, configure
   the bastion fully so that nmap-ncat comes from the correct minor
   version (minor version config is done before any package installs)
 - Configure NRPE for the bastion as well during provisioning
 - Install python-pip already in the base role instead of relying on it
   being installed by the NRPE roles (base must now run before NRPE
   roles)

Automatic updates are now disabled in the base role, so this setting
should be picked up by all hosts instead of just the ones that run
Docker.