Fix grep return code in namespace health check
When grepping for the existence of a namespace in the function that
checks for namespace health, add a "|| true" to ensure a zero return
code even if a namespace doesn't exist. This is intended behaviour in
this case, and we don't want the test to fail just because a namespace
is not found. We want to skip the test instead.