Remove remains of nodes no longer in cluster
Amend the node scaledown playbook so that it can also remove remnants of
node configuration even if a node has been deleted from the cluster
already with "oc node delete" or similar. This means we can clean up
/etc/hosts entries and generated configs for nodes even if we've
previously cleaned up the node from the cluster.

The logic of the change is that we look at the nodes we have in the
cluster and the nodes in the removal list and create an intersection of
these two lists to get a list of nodes to drain. This way only nodes
that can be drained are drained while nodes that are not in the cluster
can still be cleaned up properly.