Scaledown playbook: clarify README.md instructions
Previously the instructions didn't say anything about how you should
edit the list of nodes to remove from the cluster. Since we are in a
deployment container, this might not be obvious, as only vi is available
as an editor. Add instructions for putting the list of nodes to remove
in a file under /tmp on the deployment container using vi to edit it.

Also unify the code blocks in README.md so that they all use backticks.
This gives us syntax highlighting and the ability to give the markdown
processor a hint about how to do the syntax highlighting (i.e. which
language is the text in).