Clarify single VM recovery instructions
The single VM recovery instructions had an example rebuild command where
the image was specified as a specific CentOS image. In most cases we
don't want to specify the image but just want to use whatever image was
originally used to build the server. Thus, leave the --image option out
of the command as it is potentially dangerous. Explain what --image does
instead.

Also add a note about nova evacuate and additional advice about not
deleting servers because it messes up Heat.