Make yum repo locking optional
Add a new boolean setting lock_os_minor_version that defaults to false
and determines whether we use the latest packages from the major version
(e.g. 7) or the latest packages from the given minor version (e.g. 7.4).
Hosts can be switched between these two modes by changing the value of
the boolean.

Because mirror.centos.org only has packages for the latest minor version
available, switch to using vault.centos.org when we lock to a specific
minor release version.

Templatize the default CentOS base repo file so that we always have it
and just enable or disable the repos in it as needed.