Explicitly run keepalived scripts as root
Set the script_user setting for keepalived as "root root". If this is
not set, keepalived should default to using the root user, but this is
apparently not the case, according to the logs:

WARNING - default user 'keepalived_script' for script execution does not
exist - please create.
Unable to set default user for track script
/usr/local/share/keepalived/check_app_routes.bash - removing
  check_app_routes no match, ignoring...

If one does create the keepalived_script user and set keepalived to use
this, this still doesn't work (possibly because of SELinux). Then we get
this:

Couldn't setgid: 1001 (Operation not permitted)

If we do set script_user as "root root" then the scripts run
succesfully. Running the scripts as root should be OK in our case as no
part of the path to the script nor the script itself is writable by
non-root users.