Make post_install.yml work when called on its own
We refer to some variables for service broker install flags in
post_install.yml that will not have been set if we only run
post_install.yml itself. Add a conditional for checking whether the
variables in question are set to avoid running into an error and to
ensure post_install.yml works on its own as well.