Add InfluxDB for metrics storage
Deploy an InfluxDB instance to the monitoring-infra namespace for
long-term storage of Prometheus metrics. Configure Prometheus to write
to a remote URL to make this work. There is a remote storage adapter
listening for data in this remote URL and then forwarding that to
InfluxDB. Create a database called "prometheus" in InfluxDB via
its API and point the remote storage adapter to write to this database.

Add a Dockerfile for the Prometheus remote storage adapter that is used
for writing data from Prometheus to InfluxDB. Add this adapter as a
sidecar container to the InfluxDB deployment.

Create objects to deploy InfluxDB under templates/influxdb:

 * Deployment - this contains an InfluxDB container and the adapter
   container
 * PersistentVolume - defines a hostPath volume to be claimed by a PVC
 * PersistentVolumeClaim - this is used for storage by InfluxDB
 * Service
 * ConfigMap - this is used for configuring InfluxDB
 * Route - creates a route for accessing InfluxDB from external systems
 * SecurityContextConstraints - makes it possible to use hostPath and a
   specific UID for InfluxDB
 * ServiceAccount - used to access the correct "hostpath" SCC

Also included is a CronJob that takes nightly (by default, can be
configured) backups of the contents of InfluxDB on to the InfluxDB node.
The backup playbook is modified to pull these to the bastion as well.

The containers in the InfluxDB pod have liveness and readiness probes
configured for them. InfluxDB has a dedicated /ping API for this
purpose.

Modify the upsert task so that the object YAMLs it creates have a YAML
extension instead of j2 for Jinja templates. These are instantiated, so
they shouldn't have the j2 extension.

Setup authentication for InfluxDB so that there is an admin user account
that is used for creating a DB and writing to it and a separate reader
user account that is used by Grafana to read data from InfluxDB.

The password for the account used by Grafana to read data from InfluxDB
needs to be set in this variable:
 * influxdb_prom_reader_password

Introduce these configuration variables in the Jinja templates and the
Ansible code:

 * prometheus_remote_storage_adapter_image
 * influxdb_image
 * influxdb_req_mem
 * influxdb_req_cpu
 * influxdb_lim_mem
 * influxdb_lim_cpu
 * influxdb_prom_db_default_rp
 * influxdb_backup_cronjob_schedule
 * influxdb_vol_size
 * influxdb_backup_vol_size

These all have defaults set so the code should work even if they are not
explicitly set.

Requests and limits are defined for the InfluxDB pod. These have
defaults but can also be set using variables listed above.

A route is configured for accessing InfluxDB from external systems. The
same whitelist is used for this route as for other things that should be
accessible by monitoring hosts.

InfluxDB is set to run on a special new type of compute node with fast
local storage. It is set to use a hostPath volume. For this to work,
some modification is necessary to the security context of the InfluxDB
pod and the file access rights on the node on which it runs. We run
InfluxDB with a specific user ID (1234) and then ensure the hostPath it
is accessing is writable/readable by that UID.

Change the version locking code for CentOS so that extra repository
config files are removed when lock_os_minor_version is set to false if
they were previously added when lock_os_minor_version was true.

Add BATS tests to run against InfluxDB. Ensure the route works, auth is
enabled and queries can be executed.

Add recovery instructions for the InfluxDB machine. Make some
modifications so that the restore process works:
 - Put cloud-user into the influxdb group on the InfluxDB node so that
   it can copy backups there
 - Make creation of the prometheus DB optional during monitoring
   deployment so that we can restore it from backups instead

The restore process uses a Kubernetes Job. This Job is created by a new
restore_influxdb playbook.

Also modify the backup CronJob for InfluxDB so that files older than 45
minutes are deleted from the backup directory. This means the backup has
about 45 minutes to finish, which should be plenty of time. This can be
adjusted later on if needed.