make any ansible errors fatal
Continuing deployment for remaining hosts can lead to undetected errors.
Here we set ansible to quit execution as soon as one task fails. There
has been some turbulence on whether setting 'any_errors_fatal'
works or not, and an additional 'max_fail_percentage' is in place as a
fallback procedure.