InfluxDB: Support additional configuration queries
Make it possible to run additional queries on InfluxDB after basic
configuration. This makes it possible to set up custom retention
policies and continuous queries for downsampling in a flexible manner.
The queries can be specified as a separate list.

Because downsampling is now configured using custom queries, reduce the
default for the default retention policy to 1d instead of 7d. Longer
term data can now be configured to be stored at a lower resolution in
other databases.

Add two examples of Prometheus rules that are calculated before storing
the data. These can be used to store the result of complex Prometheus
queries in a way that can then be read using InfluxDB queries. The
examples added are cpu_consumed_per_node and cpu_consumed_per_node_type.
These pieces of information would not be possible to query via InfluxDB
otherwise based on plain Prometheus data. Other similar metrics can be
added based on the example shown.

This change makes it possible to configure downsampling for InfluxDB so
that it can be used with Grafana.