Add InfluxDB Grafana dashboard
Add an example Grafana dashboard that uses InfluxDB data. This dashboard
is added to Grafana during post-install via the HTTP dashboard API. It
comes with variable settings for configuring the datasource and data
interval to switch between different retention policies and data
resolutions.

This dashboard is intended to be a starting point and not a complete
capacity planning dashboard. One crucial thing still missing is CPU and
memory requests. These would need the kube-state-metrics add-on for
Prometheus (https://github.com/kubernetes/kube-state-metrics).